Messages with user, date, and message
Added queries need to insert messages into the db and bring them out to be shown on wall.html.
Added a variable to wall.html that holds a for loop that loops through all messages in the messages table. Then links the current user logged in to the message being posted w/ the date and the message below them.  Added in a comment form that will attach comments to each message.
Logout button was added to test all users.